requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
astroid==3.3.11
attrs==25.4.0
bleach==6.3.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
croniter==6.0.0
crontab==1.0.5
dill==0.4.0
dnspython==2.8.0
dodgy==0.2.1
flake8==7.3.0
flake8-polyfill==1.0.2
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-RQ2==18.3
freezegun==1.5.5
frozenlist==1.8.0
gunicorn==23.0.0
honcho==2.0.0
idna==3.11
isort==6.1.0
itsdangerous==2.2.0
jigsawstack==0.3.8
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
pep8-naming==0.10.0
platformdirs==4.5.0
propcache==0.4.1
prospector==1.17.3
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
pylint==3.3.9
pylint-celery==0.3
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
ratelimit==2.2.1
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
requirements-detector==1.4.0
rq==2.6.0
rq-scheduler==0.14.0
semver==3.0.4
setoptconf-tmp==0.3.1
schedule==1.2.1
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.5.0
waitress==3.0.2
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.22.0
